Windows: Disable OpenMP
The OpenMP implementation of most of the Conda-Forge software stack
is using MSVC's OpenMP runtime (limited to OpenMP 2.0). This is
also what is used in blosc in our dependencies.

When we compile WarpX, we need a newer OpenMP (4.0+), which is
why we compiled with LLVM. Unfortunately, the LLVM OpenMP runtime
and MSVC runtime cannot be loaded at the same time and thus crash
if we perform I/O operations.

For now, we just disable OpenMP for WarpX builds. In the future,
Microsoft might upgrade their OpenMP runtime:
https://devblogs.microsoft.com/cppblog/improved-openmp-support-for-cpp-in-visual-studio/